add service deploy logic
