+++ pkg/microservice/aslan/core/common/repository/models/render_set.go
@@ -70,3 +70,11 @@ func (m *RenderSet) GetServiceRenderMap() map[string]*templatemodels.ServiceRend
 	}
 	return serviceRenderMap
 }
+
+func (m *RenderSet) GetChartRenderMap() map[string]*templatemodels.ServiceRender {
+	serviceRenderMap := make(map[string]*templatemodels.ServiceRender)
+	for _, render := range m.ChartInfos {
+		serviceRenderMap[render.ServiceName] = render
+	}
+	return serviceRenderMap
+}
